Interface_DI-main/requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.9
numpy>=1.24,<2.0
Pillow>=10.0.0